feat(YAGHM-1): Do refactoring
