+++ yaghm/plugin/src/main/kotlin/dev/yaghm/plugin/internal/config/GitHookConfig.kt
@@ -1,8 +1,8 @@
 package dev.yaghm.plugin.internal.config
 
-@DslMarker
-annotation class GitHookDSL
-
+import dev.yaghm.plugin.internal.dsl.GitHookAction
+import dev.yaghm.plugin.internal.dsl.GitHookType
+import dev.yaghm.plugin.internal.dsl.preCommit
 
 class GitHookConfig {
     var type: GitHookType? = null
@@ -10,33 +10,6 @@ class GitHookConfig {
     var action: ((GitHookAction) -> Unit)? = null
     var doLastAction: ((GitHookAction) -> Unit)? = null
 }
-
-
-class GitHookAction {
-    var command: String? = null
-}
-
-
-fun GitHookConfig.action(configure: (GitHookAction) -> Unit) {
-    action = configure
-}
-
-//
-fun preCommit(configure: GitHookConfig.() -> Unit): GitHookConfig {
-    return configure("preCommit", configure)
-}
-
-fun prePush(filePath: String): GitHookConfig {
-    return GitHookConfig().apply {
-        this.type = GitHookType.PREPUSH
-        this.action = {
-            GitHookAction().apply {
-                command = filePath
-            }
-        }
-    }
-}
-//
 
 fun configure(type: String, configure: GitHookConfig.() -> Unit): GitHookConfig {
     val gitHookConfig = GitHookConfig().apply {
@@ -46,21 +19,17 @@ fun configure(type: String, configure: GitHookConfig.() -> Unit): GitHookConfig 
     return gitHookConfig
 }
 
+
 fun foo() {
     configure("preCommit") {
-        action {
+        action = {
 
         }
     }
     preCommit {
-        action {
+        action = {
 
         }
     }
 }
 
-enum class GitHookType(val type: String?) {
-    PRECOMMIT("precommit"),
-    PREPUSH("prepush"),
-    NI(null)
-}
