+++ yaghm/plugin/src/main/kotlin/dev/yaghm/plugin/internal/dsl/GitHookAction.kt
@@ -1,0 +1,11 @@
+package dev.yaghm.plugin.internal.dsl
+
+import dev.yaghm.plugin.internal.config.GitHookConfig
+
+class GitHookAction {
+    var command: String? = null
+}
+
+fun GitHookConfig.action(configure: (GitHookAction) -> Unit) {
+    action = configure
+}
