+++ yaghm/plugin/src/main/kotlin/dev/yaghm/plugin/internal/dsl/README.md
@@ -1,0 +1,1 @@
+# YAGHM
